+++ src/main/java/by/smelova/dentalclinic/rest/VisitController.java
@@ -6,12 +6,18 @@ import by.smelova.dentalclinic.models.Visit;
 import by.smelova.dentalclinic.service.DoctorService;
 import by.smelova.dentalclinic.service.PatientService;
 import by.smelova.dentalclinic.service.VisitService;
+import io.swagger.v3.oas.annotations.responses.ApiResponse;
+import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.Date;
+import java.util.List;
+
 @RestController
-@RequestMapping(value = "/api/user/")
+@RequestMapping(value = "/api/user/visit")
 public class VisitController {
     @Autowired
     private VisitService visitService;
@@ -20,14 +26,31 @@ public class VisitController {
     @Autowired
     private DoctorService doctorService;
 
-    @GetMapping(value = "/visits")
+    @GetMapping(value = "/all")
     public ResponseEntity GetVisits() {
         return ResponseEntity.ok(visitService.getAllVisits());
     }
-    @GetMapping(value = "/visit")
-    public ResponseEntity GetVisit(@RequestParam int id) {return ResponseEntity.ok(visitService.getById((long) id));  }
+    @GetMapping(value = "/byid")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200"),
+            @ApiResponse(responseCode = "404")})
+    public ResponseEntity GetVisit(@RequestParam int id) {
+        Visit rc = visitService.getById((long) id);
+        if(rc != null)
+        return ResponseEntity.ok(rc);
+        else return ResponseEntity.status(404).build();
+    }
+    @GetMapping(value = "/bydate")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200"),
+            @ApiResponse(responseCode = "404")})
+    public ResponseEntity GetVisitByDate(@RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) Date date) {
+        List<Visit> rc = visitService.getByDate(date);
+        if(rc.size() !=0) return ResponseEntity.ok(rc);
+        else return ResponseEntity.status(404).build();
+    }
 
-    @PostMapping(value = "/visit/add")
+    @PostMapping(value = "/add")
     public ResponseEntity AddVisit(@RequestBody VisitDto visitdto) {
         Visit visit = Mapper.map(visitdto, Visit.class);
         visit.setDoctor(doctorService.getById(visitdto.getDoctorID()));
@@ -36,16 +59,22 @@ public class VisitController {
         return ResponseEntity.ok(visit);
     }
 
-    @PutMapping(value = "/visit/edit")
+    @PutMapping(value = "/edit")
     public ResponseEntity EditVisit(@RequestBody VisitDto visitDto) {
         return ResponseEntity.ok(visitService.EditVisit(visitDto));
     }
 
-    @DeleteMapping(value = "/visit/delete/{id}")
+    @DeleteMapping(value = "/delete/{id}")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200"),
+            @ApiResponse(responseCode = "404")})
     public ResponseEntity DeleteVisit(@PathVariable Long id) {
         Visit rc = visitService.getById(id);
-        visitService.deleteById(Math.toIntExact(id));
-        return ResponseEntity.ok(rc);
+        if(rc != null) {
+            visitService.deleteById(Math.toIntExact(id));
+            return ResponseEntity.ok(rc);
+        } else return ResponseEntity.status(404).build();
+
     }
 
 }
